backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
sqlmodel>=0.0.8
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.24
